+++ api/src/team/application/certification-center-invitation-controller.js
@@ -1,5 +1,5 @@
-import { usecases } from '../../../lib/domain/usecases/index.js';
 import { extractLocaleFromRequest } from '../../../lib/infrastructure/utils/request-response-utils.js';
+import { usecases } from '../domain/usecases/index.js';
 
 const sendInvitations = async function (request, h) {
   const certificationCenterId = request.params.certificationCenterId;
